+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.8
+
+Fix to some of the Firelands Normal tier tokens
+
 # 0.0.7
 
 Fixes to some Ulduar and Throne of Thunder tier tokens, added Undermine Raid Tokens, added Nerubar Warbound Tokens